[ML] Refactor ProcessCtrl into Autodetect and Normalizer builders
This moves the helper functionality for creating the autodetect
and mormalizer processes into corresponding builders.
