+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/job/process/autodetect/AutodetectBuilder.java
@@ -6,51 +6,124 @@
 package org.elasticsearch.xpack.ml.job.process.autodetect;
 
 import org.apache.logging.log4j.Logger;
+import org.elasticsearch.common.Strings;
+import org.elasticsearch.common.settings.Setting;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
+import org.elasticsearch.xpack.core.XPackPlugin;
 import org.elasticsearch.xpack.core.ml.calendars.ScheduledEvent;
+import org.elasticsearch.xpack.core.ml.job.config.AnalysisConfig;
 import org.elasticsearch.xpack.core.ml.job.config.AnalysisLimits;
+import org.elasticsearch.xpack.core.ml.job.config.DataDescription;
 import org.elasticsearch.xpack.core.ml.job.config.Job;
 import org.elasticsearch.xpack.core.ml.job.config.MlFilter;
 import org.elasticsearch.xpack.core.ml.job.config.ModelPlotConfig;
-import org.elasticsearch.xpack.ml.job.process.NativeController;
-import org.elasticsearch.xpack.ml.job.process.ProcessCtrl;
-import org.elasticsearch.xpack.ml.job.process.ProcessPipes;
 import org.elasticsearch.xpack.core.ml.job.process.autodetect.state.Quantiles;
+import org.elasticsearch.xpack.ml.job.process.NativeController;
+import org.elasticsearch.xpack.ml.job.process.ProcessBuilderUtils;
+import org.elasticsearch.xpack.ml.job.process.ProcessPipes;
 import org.elasticsearch.xpack.ml.job.process.autodetect.writer.AnalysisLimitsWriter;
 import org.elasticsearch.xpack.ml.job.process.autodetect.writer.FieldConfigWriter;
 import org.elasticsearch.xpack.ml.job.process.autodetect.writer.ModelPlotConfigWriter;
 
+import java.io.BufferedWriter;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
+import java.util.Random;
 import java.util.Set;
+
+import static org.elasticsearch.xpack.ml.job.process.ProcessBuilderUtils.addIfNotNull;
 
 /**
  * The autodetect process builder.
  */
 public class AutodetectBuilder {
+
+    /**
+     * Autodetect API native program name - always loaded from the same directory as the controller process
+     */
+    public static final String AUTODETECT = "autodetect";
+    static final String AUTODETECT_PATH = "./" + AUTODETECT;
+
+    /*
+     * Arguments used by both autodetect and normalize
+     */
+    public static final String BUCKET_SPAN_ARG = "--bucketspan=";
+    public static final String DELETE_STATE_FILES_ARG = "--deleteStateFiles";
+    public static final String LENGTH_ENCODED_INPUT_ARG = "--lengthEncodedInput";
+    public static final String MODEL_CONFIG_ARG = "--modelconfig=";
+    public static final String QUANTILES_STATE_PATH_ARG = "--quantilesState=";
+    public static final String PER_PARTITION_NORMALIZATION = "--perPartitionNormalization";
+
     private static final String CONF_EXTENSION = ".conf";
+    static final String JOB_ID_ARG = "--jobid=";
     private static final String LIMIT_CONFIG_ARG = "--limitconfig=";
     private static final String MODEL_PLOT_CONFIG_ARG = "--modelplotconfig=";
     private static final String FIELD_CONFIG_ARG = "--fieldconfig=";
+    static final String LATENCY_ARG = "--latency=";
+    static final String RESULT_FINALIZATION_WINDOW_ARG = "--resultFinalizationWindow=";
+    static final String MULTIVARIATE_BY_FIELDS_ARG = "--multivariateByFields";
+    static final String PERSIST_INTERVAL_ARG = "--persistInterval=";
+    static final String MAX_QUANTILE_INTERVAL_ARG = "--maxQuantileInterval=";
+    static final String SUMMARY_COUNT_FIELD_ARG = "--summarycountfield=";
+    static final String TIME_FIELD_ARG = "--timefield=";
 
-    private Job job;
-    private List<Path> filesToDelete;
-    private Logger logger;
+    /**
+     * Name of the config setting containing the path to the logs directory
+     */
+    private static final int DEFAULT_MAX_NUM_RECORDS = 500;
+
+    /**
+     * The maximum number of anomaly records that will be written each bucket
+     */
+    public static final Setting<Integer> MAX_ANOMALY_RECORDS_SETTING = Setting.intSetting("max.anomaly.records", DEFAULT_MAX_NUM_RECORDS,
+            Setting.Property.NodeScope);
+
+    /**
+     * Config setting storing the flag that disables model persistence
+     */
+    public static final Setting<Boolean> DONT_PERSIST_MODEL_STATE_SETTING = Setting.boolSetting("no.model.state.persist", false,
+            Setting.Property.NodeScope);
+
+    private static final int SECONDS_IN_HOUR = 3600;
+
+    /**
+     * Roughly how often should the C++ process persist state?  A staggering
+     * factor that varies by job is added to this.
+     */
+    private static final long DEFAULT_BASE_PERSIST_INTERVAL = 10800; // 3 hours
+
+    /**
+     * Roughly how often should the C++ process output quantiles when no
+     * anomalies are being detected?  A staggering factor that varies by job is
+     * added to this.
+     */
+    static final int BASE_MAX_QUANTILE_INTERVAL = 21600; // 6 hours
+
+    /**
+     * Persisted quantiles are written to disk so they can be read by
+     * the autodetect program.  All quantiles files have this extension.
+     */
+    private static final String QUANTILES_FILE_EXTENSION = ".json";
+
+    private final Job job;
+    private final List<Path> filesToDelete;
+    private final Logger logger;
+    private final Environment env;
+    private final Settings settings;
+    private final NativeController controller;
+    private final ProcessPipes processPipes;
     private Set<MlFilter> referencedFilters;
     private List<ScheduledEvent> scheduledEvents;
     private Quantiles quantiles;
-    private Environment env;
-    private Settings settings;
-    private NativeController controller;
-    private ProcessPipes processPipes;
 
     /**
      * Constructs an autodetect process builder
@@ -98,7 +171,7 @@ public class AutodetectBuilder {
      */
     public void build() throws IOException {
 
-        List<String> command = ProcessCtrl.buildAutodetectCommand(env, settings, job, logger);
+        List<String> command = buildAutodetectCommand();
 
         buildLimits(command);
         buildModelPlotConfig(command);
@@ -107,6 +180,101 @@ public class AutodetectBuilder {
         buildFieldConfig(command);
         processPipes.addArgs(command);
         controller.startProcess(command);
+    }
+
+    /**
+     * Visible for testing
+     */
+    List<String> buildAutodetectCommand() {
+        List<String> command = new ArrayList<>();
+        command.add(AUTODETECT_PATH);
+
+        String jobId = JOB_ID_ARG + job.getId();
+        command.add(jobId);
+
+        AnalysisConfig analysisConfig = job.getAnalysisConfig();
+        if (analysisConfig != null) {
+            addIfNotNull(analysisConfig.getBucketSpan(), BUCKET_SPAN_ARG, command);
+            addIfNotNull(analysisConfig.getLatency(), LATENCY_ARG, command);
+            addIfNotNull(analysisConfig.getSummaryCountFieldName(),
+                    SUMMARY_COUNT_FIELD_ARG, command);
+            if (Boolean.TRUE.equals(analysisConfig.getOverlappingBuckets())) {
+                Long window = analysisConfig.getResultFinalizationWindow();
+                if (window == null) {
+                    window = AnalysisConfig.DEFAULT_RESULT_FINALIZATION_WINDOW;
+                }
+                command.add(RESULT_FINALIZATION_WINDOW_ARG + window);
+            }
+            if (Boolean.TRUE.equals(analysisConfig.getMultivariateByFields())) {
+                command.add(MULTIVARIATE_BY_FIELDS_ARG);
+            }
+
+            if (analysisConfig.getUsePerPartitionNormalization()) {
+                command.add(PER_PARTITION_NORMALIZATION);
+            }
+        }
+
+        // Input is always length encoded
+        command.add(LENGTH_ENCODED_INPUT_ARG);
+
+        // Limit the number of output records
+        command.add(maxAnomalyRecordsArg(settings));
+
+        // always set the time field
+        String timeFieldArg = TIME_FIELD_ARG + getTimeFieldOrDefault(job);
+        command.add(timeFieldArg);
+
+        int intervalStagger = calculateStaggeringInterval(job.getId());
+        logger.debug("Periodic operations staggered by " + intervalStagger +" seconds for job '" + job.getId() + "'");
+
+        // Supply a URL for persisting/restoring model state unless model
+        // persistence has been explicitly disabled.
+        if (DONT_PERSIST_MODEL_STATE_SETTING.get(settings)) {
+            logger.info("Will not persist model state - "  + DONT_PERSIST_MODEL_STATE_SETTING + " setting was set");
+        } else {
+            // Persist model state every few hours even if the job isn't closed
+            long persistInterval = (job.getBackgroundPersistInterval() == null) ?
+                    (DEFAULT_BASE_PERSIST_INTERVAL + intervalStagger) :
+                    job.getBackgroundPersistInterval().getSeconds();
+            command.add(PERSIST_INTERVAL_ARG + persistInterval);
+        }
+
+        int maxQuantileInterval = BASE_MAX_QUANTILE_INTERVAL + intervalStagger;
+        command.add(MAX_QUANTILE_INTERVAL_ARG + maxQuantileInterval);
+
+        if (ProcessBuilderUtils.modelConfigFilePresent(env)) {
+            String modelConfigFile = XPackPlugin.resolveConfigFile(env, ProcessBuilderUtils.ML_MODEL_CONF).toString();
+            command.add(MODEL_CONFIG_ARG + modelConfigFile);
+        }
+
+        return command;
+    }
+
+
+    static String maxAnomalyRecordsArg(Settings settings) {
+        return "--maxAnomalyRecords=" + MAX_ANOMALY_RECORDS_SETTING.get(settings);
+    }
+
+    private static String getTimeFieldOrDefault(Job job) {
+        DataDescription dataDescription = job.getDataDescription();
+        boolean useDefault = dataDescription == null
+                || Strings.isNullOrEmpty(dataDescription.getTimeField());
+        return useDefault ? DataDescription.DEFAULT_TIME_FIELD : dataDescription.getTimeField();
+    }
+
+    /**
+     * This random time of up to 1 hour is added to intervals at which we
+     * tell the C++ process to perform periodic operations.  This means that
+     * when there are many jobs there is a certain amount of staggering of
+     * their periodic operations.  A given job will always be given the same
+     * staggering interval (for a given JVM implementation).
+     *
+     * @param jobId The ID of the job to calculate the staggering interval for
+     * @return The staggering interval
+     */
+    static int calculateStaggeringInterval(String jobId) {
+        Random rng = new Random(jobId.hashCode());
+        return rng.nextInt(SECONDS_IN_HOUR);
     }
 
     private void buildLimits(List<String> command) throws IOException {
@@ -152,13 +320,30 @@ public class AutodetectBuilder {
         if (quantiles != null && !quantiles.getQuantileState().isEmpty()) {
             logger.info("Restoring quantiles for job '" + job.getId() + "'");
 
-            Path normalizersStateFilePath = ProcessCtrl.writeNormalizerInitState(
-                    job.getId(), quantiles.getQuantileState(), env);
+            Path normalizersStateFilePath = writeNormalizerInitState(job.getId(), quantiles.getQuantileState(), env);
 
-            String quantilesStateFileArg = ProcessCtrl.QUANTILES_STATE_PATH_ARG + normalizersStateFilePath;
+            String quantilesStateFileArg = QUANTILES_STATE_PATH_ARG + normalizersStateFilePath;
             command.add(quantilesStateFileArg);
-            command.add(ProcessCtrl.DELETE_STATE_FILES_ARG);
+            command.add(DELETE_STATE_FILES_ARG);
         }
+    }
+
+    /**
+     * Write the normalizer init state to file.
+     */
+    public static Path writeNormalizerInitState(String jobId, String state, Environment env)
+            throws IOException {
+        // createTempFile has a race condition where it may return the same
+        // temporary file name to different threads if called simultaneously
+        // from multiple threads, hence add the thread ID to avoid this
+        Path stateFile = Files.createTempFile(env.tmpFile(), jobId + "_quantiles_" + Thread.currentThread().getId(),
+                QUANTILES_FILE_EXTENSION);
+
+        try (BufferedWriter osw = Files.newBufferedWriter(stateFile, StandardCharsets.UTF_8)) {
+            osw.write(state);
+        }
+
+        return stateFile;
     }
 
     private void buildFieldConfig(List<String> command) throws IOException {
